+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added the `logging-level` option (see `config.yml` for more details)
 - Added some more translation strings under `team.inventory` and `team.colors` in the `lang.yml` file
+- Added `players.death-message` to `lang.yml`
 
 ### Bug fixes